aider: Fixed incorrect assignment of array to `threshold` and `S_threshold` variables.
# Aider chat conversation:

USER: Error at 279:5 Variable 'threshold' was declared with 'series float' type. Cannot assign it expression of type 'float[]'

ASSISTANT: It seems there is an error in the code where a `float[]` type is being assigned to a variable `threshold` that is expected to be of type `series float`. To fix this, we need to ensure that we are assigning the correct type to the `threshold` variable. I will provide a diff to correct this issue:

```diff
--- TTB_7_6_backup.pine
+++ TTB_7_6_backup.pine
@@ ... @@
 if strategy.position_size > 0 and i == 0
     atrp_distance = atrp_daily * max_atrp_SO_distance
     var float[] SO_distance = array.new_float(max_safe_order)
     for j = 1 to max_safe_order
         if j == 1
             array.set(SO_distance, j - 1, initial_order - initial_order * price_deviation)
         else
             array.set(SO_distance, j - 1, initial_order - initial_order * (price_deviation + (atrp_distance - price_deviation) * (math.pow(j / max_safe_order, 2))))
-    threshold := array.get(SO_distance, current_so - 1)
-    threshold := SO_distance[current_so]
+    threshold := array.get(SO_distance, current_so - 1)
 else if strategy.position_size < 0 and i == 0
     atrp_distance = atrp_daily * max_atrp_SO_distance
     var float[] S_SO_distance = array.new_float(max_S_safe_order)
     for j = 1 to max_S_safe_order
         if j == 1
             array.set(S_SO_distance, j - 1, initial_S_order * (1 + price_S_deviation))
         else
             array.set(S_SO_distance, j - 1, initial_S_order * (1 + price_S_deviation + (atrp_distance - price_S_deviation) * (math.pow(j / max_S_safe_order, 2))))
-    S_threshold := array.get(S_SO_distance, current_so - 1)
-    S_threshold := S_SO_distance[current_so]
+    S_threshold := array.get(S_SO_distance, current_so - 1)
```

This diff removes the incorrect assignment of the entire array to the `threshold` variable and ensures that only the specific element from the array is assigned to it. Apply this diff to the `TTB_7_6_backup.pine` file to fix the error.